+++ torchmetrics/functional/regression/mean_absolute_error.py
@@ -42,12 +42,11 @@ def mean_absolute_error(preds: Tensor, target: Tensor) -> Tensor:
         Tensor with MAE
 
     Example:
-
+        >>> from torchmetrics.functional import mean_absolute_error
         >>> x = torch.tensor([0., 1, 2, 3])
         >>> y = torch.tensor([0., 1, 2, 2])
         >>> mean_absolute_error(x, y)
         tensor(0.2500)
-
     """
     sum_abs_error, n_obs = _mean_absolute_error_update(preds, target)
     return _mean_absolute_error_compute(sum_abs_error, n_obs)